stock_service/requirements.txt
cassandra-driver==3.23.0
click==7.1.2
Flask==1.1.2
geomet==0.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
six==1.14.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
gunicorn==20.0.4
simplejson==3.17.0